+++ common/text_opinions/helper.py
@@ -1,9 +1,6 @@
 from arekit.common.entities.base import Entity
-from arekit.common.parsed_news.base import ParsedNews
 from arekit.common.parsed_news.collection import ParsedNewsCollection
 from arekit.common.parsed_news.term_position import TermPosition, TermPositionTypes
-from arekit.common.synonyms import SynonymsCollection
-from arekit.common.text_frame_variant import TextFrameVariant
 from arekit.common.text_opinions.collection import TextOpinionCollection
 from arekit.common.text_opinions.text_opinion import TextOpinion
 from arekit.common.text_opinions.end_type import EntityEndType
@@ -22,6 +19,10 @@ class TextOpinionHelper(object):
         parsed news, positions, text_opinions
     """
 
+    def __init__(self, parsed_news_collection):
+        assert(isinstance(parsed_news_collection, ParsedNewsCollection))
+        self.__parsed_news_collection = parsed_news_collection
+
     # region public 'extract' methods
 
     @staticmethod
@@ -29,69 +30,53 @@ class TextOpinionHelper(object):
         parsed_news, id = TextOpinionHelper.__get(text_opinion, end_type)
         return parsed_news.get_entity_value(id)
 
+    def extract_entity_position(self, text_opinion, end_type, position_type=None):
+        return self.__get_entity_position_2(text_opinion=text_opinion,
+                                            end_type=end_type,
+                                            position_type=position_type)
+
     @staticmethod
     def extract_entity_sentence_index(text_opinion, end_type):
         entity_position = TextOpinionHelper.__get_entity_position(text_opinion, end_type)
         return entity_position.get_index(TermPositionTypes.SentenceIndex)
 
-    @staticmethod
-    def extract_entity_sentence_level_term_index(text_opinion, end_type):
-        parsed_news, id = TextOpinionHelper.__get(text_opinion, end_type)
-        term_position = parsed_news.get_entity_position(id)
-        return term_position.get_index(TermPositionTypes.IndexInSentence)
-
-    @staticmethod
-    def extract_entity_sentence_level_synonym_indices(text_opinion, end_type, synonyms):
-        assert(isinstance(synonyms, SynonymsCollection))
-        parsed_news, id = TextOpinionHelper.__get(text_opinion, end_type)
-
-        entity_pos = parsed_news.get_entity_position(id)
-        s_index = entity_pos.get_index(TermPositionTypes.SentenceIndex)
-
-        e_value = parsed_news.get_entity_value(id)
-        e_group = synonyms.get_synonym_group_index(e_value)
-
-        inds = []
-
-        it = parsed_news.iter_sentence_terms(
-            sentence_index=s_index,
-            term_check=lambda term: isinstance(term, Entity))
-
-        for e_index, e in it:
-
-            if not synonyms.contains_synonym_value(e.Value):
-                if e_value != e.Value:
-                    continue
-            elif e_group != synonyms.get_synonym_group_index(e.Value):
-                continue
-
-            inds.append(e_index)
-
-        return inds
-
     # endregion
 
     # region public 'calculate' methods
 
-    @staticmethod
-    def calculate_distance_between_text_opinion_ends_in_terms(text_opinion):
+    def calculate_distance_between_text_opinion_ends_in_terms(self, text_opinion):
         assert(isinstance(text_opinion, TextOpinion))
 
-        pos1 = TextOpinionHelper.__get_entity_position(text_opinion=text_opinion,
-                                                       end_type=EntityEndType.Source)
+        parsed_news = self.__parsed_news_collection.get_by_news_id(text_opinion)
 
-        pos2 = TextOpinionHelper.__get_entity_position(text_opinion=text_opinion,
-                                                       end_type=EntityEndType.Target)
+        e1_id = self.__get_end_id(text_opinion=text_opinion,
+                                  end_type=EntityEndType.Source)
 
-        return TextOpinionHelper.__calc_distance_in_terms(pos1=pos1, pos2=pos2)
+        e2_id = self.__get_end_id(text_opinion=text_opinion,
+                                  end_type=EntityEndType.Target)
 
-    @staticmethod
-    def calculate_distance_between_entities_in_terms(parsed_news, e1, e2):
-        assert(isinstance(parsed_news, ParsedNews))
+        return TextOpinionHelper.__calc_distance(
+            pos1=parsed_news.get_entity_position(id_in_document=e1_id),
+            pos2=parsed_news.get_entity_position(id_in_document=e2_id))
+
+    def calculate_distance_between_entities_in_sentences(self, news_id, e1, e2):
+        assert(isinstance(news_id, int))
         assert(isinstance(e1, Entity))
         assert(isinstance(e2, Entity))
 
-        return TextOpinionHelper.__calc_distance_in_terms(
+        parsed_news = self.__parsed_news_collection.get_by_news_id(news_id)
+        return TextOpinionHelper.__calc_distance(
+            pos1=parsed_news.get_entity_position(e1.IdInDocument),
+            pos2=parsed_news.get_entity_position(e2.IdInDocument),
+            position_type=TermPositionTypes.SentenceIndex)
+
+    def calculate_distance_between_entities_in_terms(self, news_id, e1, e2):
+        assert(isinstance(news_id, int))
+        assert(isinstance(e1, Entity))
+        assert(isinstance(e2, Entity))
+
+        parsed_news = self.__parsed_news_collection.get_by_news_id(news_id)
+        return TextOpinionHelper.__calc_distance(
             pos1=parsed_news.get_entity_position(e1.IdInDocument),
             pos2=parsed_news.get_entity_position(e2.IdInDocument))
 
@@ -99,45 +84,67 @@ class TextOpinionHelper(object):
 
     # region public 'check' methods
 
-    @staticmethod
-    def check_ends_has_same_sentence_index(text_opinion):
+    def check_ends_has_same_sentence_index(self, text_opinion):
         assert(isinstance(text_opinion, TextOpinion))
 
-        pos1 = TextOpinionHelper.__get_entity_position(text_opinion, EntityEndType.Source)
-        pos2 = TextOpinionHelper.__get_entity_position(text_opinion, EntityEndType.Target)
+        pos1 = self.__get_entity_position_2(text_opinion=text_opinion,
+                                            end_type=EntityEndType.Source,
+                                            position_type=TermPositionTypes.SentenceIndex)
 
-        return pos1.get_index(TermPositionTypes.SentenceIndex) == \
-               pos2.get_index(TermPositionTypes.SentenceIndex)
+        pos2 = self.__get_entity_position_2(text_opinion=text_opinion,
+                                            end_type=EntityEndType.Target,
+                                            position_type=TermPositionTypes.SentenceIndex)
+
+        return pos1 == pos2
 
     # endregion
 
     # region public 'iter' methods
 
-    @staticmethod
-    def iter_frame_variants_with_indices_in_sentence(text_opinion):
-        parsed_news, t_id = TextOpinionHelper.__get(text_opinion, EntityEndType.Target)
-        s_index = TextOpinionHelper.extract_entity_sentence_index(text_opinion, EntityEndType.Source)
-        return parsed_news.iter_sentence_terms(
-            sentence_index=s_index,
-            term_check=lambda term: isinstance(term, TextFrameVariant))
+    def iter_terms_in_related_sentence(self, text_opinion, return_ind_in_sent, term_check=None):
+        assert(isinstance(return_ind_in_sent, bool))
+
+        id_in_doc = TextOpinionHelper.__get_end_id(text_opinion, EntityEndType.Target)
+        parsed_news = self.__parsed_news_collection.get_by_news_id(news_id=text_opinion.NewsID)
+        s_index = parsed_news.get_entity_position(id_in_document=id_in_doc,
+                                                  position_type=TermPositionTypes.SentenceIndex)
+        it = parsed_news.iter_sentence_terms(sentence_index=s_index,
+                                             term_check=term_check)
+
+        for ind_in_sent, term in it:
+            if return_ind_in_sent:
+                yield ind_in_sent, term
+            else
+                yield term
 
     # endregion
 
     # region private methods
 
+    def __get_entity_position_2(self, text_opinion, end_type, position_type=None):
+        end_id = TextOpinionHelper.__get_end_id(text_opinion, end_type)
+        parsed_news = self.__parsed_news_collection.get_by_news_id(text_opinion.NewsID)
+        return parsed_news.get_entity_position(end_id, position_type)
+
+    # TODO. Should be removed.
+    # TODO. Should be removed.
+    # TODO. Should be removed.
     @staticmethod
     def __get_entity_position(text_opinion, end_type):
         parsed_news, _id = TextOpinionHelper.__get(text_opinion, end_type)
         return parsed_news.get_entity_position(_id)
 
     @staticmethod
-    def __calc_distance_in_terms(pos1, pos2):
+    def __calc_distance(pos1, pos2, position_type=TermPositionTypes.IndexInDocument):
         assert(isinstance(pos1, TermPosition))
         assert(isinstance(pos2, TermPosition))
 
-        return abs(pos1.get_index(TermPositionTypes.IndexInDocument) -
-                   pos2.get_index(TermPositionTypes.IndexInDocument))
+        return abs(pos1.get_index(position_type) -
+                   pos2.get_index(position_type))
 
+    # TODO. Should be removed.
+    # TODO. Should be removed.
+    # TODO. Should be removed.
     @staticmethod
     def __get(text_opinion, end_type):
         owner = text_opinion.Owner